requirements.txt
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.3
joblib==1.1.1
Markdown==3.1.1
# numpy==1.24.1 numpy.bool has been deprecated as of Numpy 1.20.
numpy==1.19.5
pandas==0.25.2
requests==2.22.0
#scikit-learn==1.2.1
scikit-learn==0.24.2